rocm: change the event id type to uint64_t in backend
Preparatory commit to increase the size of the event id datatype in the
component backend layer so to make it ready for hosting event id encoded
information, such as device and instance numbers.

+++ src/components/rocm/roc_common.c
@@ -71,7 +71,7 @@ rocc_err_get_last(const char **err_string)
 }
 
 int
-rocc_dev_get_map(rocc_dev_get_map_cb query_dev_id, unsigned int *events_id, int num_events, rocc_bitmap_t *bitmap)
+rocc_dev_get_map(rocc_dev_get_map_cb query_dev_id, uint64_t *events_id, int num_events, rocc_bitmap_t *bitmap)
 {
     int i;
     rocc_bitmap_t device_map_acq = 0;
